Reset watches after reconnected.
Watches are maintained locally at the ZooKeeper server to which the
client is connected. When a client reconnects to a new server, client
need to reset all wathers if needed (according to `disable_watch_auto_reset`).

The watch events still may be missed if the znode chagned before the watcher
reset. So it is not recommended to use auto reset for applications if
you care. The better way is using monitor event to detect connection changes.

Read [ZooKeeper Watches](https://zookeeper.apache.org/doc/trunk/zookeeperProgrammers.html#ch_zkWatches)
for more.