Created event is enabled with a call to exists, but not to get_data
Even the official client implementation for the created event is enabled to
a get_data call, but actually the server won't trigger the watcher on
get_data for the created event.

Actually, if client request a get_data with watch for a znode which dose
not exist, the ZooKeeper server won't set a data watch for this znode.
So combine the ExistWatch and GetDataCreateWatch will confused users.