initialize ping_interval with infinity
This is to handle situation when zookeeper is not available when
erlzk_conn process starts. In this case ping_interval in state will be
'undefined', and the next incoming request will trigger a handle_call
callback which returns {ok, State, PingInt} to gen_server. The process
will crash in the main gen_server loop function on receive ... after
because Time will be 'undefined'.

By setting ping_interval to infinity we just emulate default behaviour.